+++ src/common/mwc/handle-request-selected-event.ts
